wip on mysql cleanup
